reverting my last commit
After doing some research and further testing, I realized was in a beta version of a future release when validating the other product types. The product team plans on keeping the original 2 products that can be assigned to a bundle. They do not have immediate plans to include other product types.